% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_iscoII.R
\name{add_iscoII}
\alias{add_iscoII}
\title{Duplicate rows to count each student two times - one for each parent's occupation - while calculating ave perf.}
\usage{
add_iscoII(isco_cols, data)
}
\arguments{
\item{isco_cols}{Names of columns with mother's and father's occupation codes - in this order.}

\item{data}{Name of a data frame containing columns given in previous arguments.}
}
\value{
Data frame given in data argument with twice as much rows and added column
        with both parents' isco codes.
}
\description{
Duplicate rows to count each student two times - one for each parent's occupation - while calculating ave perf.
}

