% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{addISCO}
\alias{addISCO}
\title{Add columns with occupations and change format to long.}
\usage{
addISCO(dataSource, NAcodes)
}
\arguments{
\item{dataSource}{data frame returned by importFromTxt function.}

\item{NAcodes}{occupation that need to be converted to NA.}
}
\value{
data frame in long format ...
}
\description{
Add columns with occupations and change format to long.
}

