% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{addSubject}
\alias{addSubject}
\title{Helper function that adds subjects to data frames in importFromTxt function.}
\usage{
addSubject(slist, subs)
}
\arguments{
\item{slist}{List containing tibbles.}

\item{subs}{Character vector of subject names.}
}
\value{
list
}
\description{
Helper function that adds subjects to data frames in importFromTxt function.
}

