% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreoccupations.R
\name{exploreOccupations}
\alias{exploreOccupations}
\title{Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.}
\usage{
exploreOccupations()
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.
}

