% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{exportSpreadsheet}
\alias{exportSpreadsheet}
\title{Export data for one year to spreadsheet file.}
\usage{
exportSpreadsheet(dataSource, outputFilePath)
}
\arguments{
\item{outputFilePath}{Path to an output file.}

\item{sourceTibble}{tibble returned by importFromTxt function.}
}
\value{
list of objects that were written to a file (invisible)
}
\description{
Export data for one year to spreadsheet file.
}

