% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givelabel.R
\name{give_label}
\alias{give_label}
\title{Labels for shiny app's tooltips.}
\usage{
give_label(point)
}
\arguments{
\item{point}{Object from a ggvis plot.}
}
\value{
html object for a tooltip.
}
\description{
Labels for shiny app's tooltips.
}

