% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givelabel.R
\name{giveLabel}
\alias{giveLabel}
\title{Labels for shiny app's tooltips.}
\usage{
giveLabel(subj, cnt, isco, avep, se, pop)
}
\arguments{
\item{subj}{Column with the subject name.}

\item{cnt}{Column with the country label.}

\item{isco}{Column with the isco label.}

\item{avep}{Column with students' average performance.}

\item{se}{Column with standard errors.}

\item{pop}{Column with population share.}
}
\value{
html object for a tooltip.
}
\description{
Labels for shiny app's tooltips.
}

