% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{importFromTxt}
\alias{importFromTxt}
\title{Import data from .txt files from OECD website.}
\usage{
importFromTxt(inputFilePath, variablesStartPositions, variablesEndPositions,
  groupingVariablesStartPositions, groupingVariablesEndPositions,
  groupingVariablesNames, studyYear, NAcodes = character(0),
  occupationNAcodes = character(0), outputFilePath = NULL)
}
\arguments{
\item{inputFilePath}{path to the file to import.}

\item{variablesStartPositions}{starting position of column containing
country code, school id, mother and father occupation
and plausible values and replicate weights.}

\item{variablesEndPositions}{ending positions of column given in previous argument.}

\item{groupingVariablesStartPositions}{starting positions of additional columns 
with factor variables.}

\item{groupingVariablesEndPositions}{ending positions of variables from previous argument.}

\item{groupingVariablesNames}{names for variables given in previous two arguments.}

\item{studyYear}{year of study.}

\item{NAcodes}{codes for NA values in PV and weight variables.}

\item{outputFilePath}{path to .rda file to which results will be saved.}
}
\value{
tibble
}
\description{
Import data from .txt files from OECD website.
}

