% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscocat.R
\name{isco_cat}
\alias{isco_cat}
\title{Extract first and first two digits of ISCO code as factors both for mother's and father's occupation.}
\usage{
isco_cat(isco_cols, data)
}
\arguments{
\item{isco_cols}{Names of columns with mother's and father's occupation codes (in this order).}

\item{data}{Name of a data frame containing columns given in previous arguments.}
}
\value{
Data frame given in 'data' argument with four described columns added.
}
\description{
Extract first and first two digits of ISCO code as factors both for mother's and father's occupation.
}

