% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpv.R
\name{mean.pv}
\alias{mean.pv}
\title{Student's average performance by given factors in selected year.}
\usage{
\method{mean}{pv}(pvlabs, groups, weights, data)
}
\arguments{
\item{pvlabs}{Names of column that contain five plausible values.}

\item{groups}{Names of the grouping variables.}

\item{weights}{Name of the column that contains final student weight.}

\item{data}{Name of a data frame containing columns given as previous arguments.}
}
\value{
Data frame containing levels of given factor variables and average performances and sums of weights
        for each combination of these factors.
}
\description{
Student's average performance by given factors in selected year.
}

