% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanby1cat.R
\name{mean_by_1_cat}
\alias{mean_by_1_cat}
\title{Average student's performances by country and one isco category.}
\usage{
mean_by_1_cat(subject, cnt_col, isco_cols, cat, final_weights, data)
}
\arguments{
\item{subject}{String of the form "MATH", "READ", "SCIE".}

\item{cnt_col}{Name of a column with country codes.}

\item{isco_cols}{Names of two columns with isco codes for mother's and father's occupation.}

\item{cat}{Character with a number of an isco category.}

\item{final_weight}{Name of a column that contains final student weights.}
}
\value{
Data frame with names of countries and means for a given isco category.
}
\description{
Average student's performances by country and one isco category.
}

