% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanbyallcats.R
\name{mean_by_all_cats}
\alias{mean_by_all_cats}
\title{Average student performance by country and ISCO category - all categories.}
\usage{
mean_by_all_cats(subject, cnt_col, isco_cols, final_weight, data)
}
\arguments{
\item{subject}{String of the form "MATH", "READ", "SCIE".}

\item{cnt_col}{Name of a column with country codes.}

\item{isco_cols}{Names of two columns with isco codes for mother's and father's occupation.}

\item{final_weight}{Name of a column that contains final student weights.}
}
\value{
Data frame with names of countries and means for isco categories.
}
\description{
Average student performance by country and ISCO category - all categories.
}

