% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpv.R
\name{mean_o}
\alias{mean_o}
\title{Mean calculated from five plausible values.}
\usage{
mean_o(means_ppv)
}
\arguments{
\item{means_ppv}{Data frame returned by mean_ppvs function that contains average performances computed
for each plausible value.}
}
\value{
Vector of student's average performances.
}
\description{
Mean calculated from five plausible values.
}

