% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanppvs.R
\name{mean_ppvs}
\alias{mean_ppvs}
\title{Mean calculated separately for each of the five plausible values and a given weight.}
\usage{
mean_ppvs(data, pvname, groups, weights = "W_FSTUWT", school_id = "SCH_ID")
}
\arguments{
\item{data}{Name of a data frame containing columns given in other arguments.}

\item{pvname}{String of the form "MATH", "READ", "SCIE".}

\item{groups}{Name of one or more factors used for grouping.}

\item{weights}{Name of a column that contains final student weights.}

\item{school_id}{Name of a column with school IDs.}
}
\value{
Data frame containing columns for each of given factor variables, five columns with
        means calculated for each plausible values and a column with sums of weights.
}
\description{
Mean calculated separately for each of the five plausible values and a given weight.
}

