% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpv.R
\name{mean_pv}
\alias{mean_pv}
\title{Average student's performances with standard errors by given factor variables.}
\usage{
mean_pv(pvname, groups, final_weight, data, se = F, brr_weights = "")
}
\arguments{
\item{pvname}{String of the form "MATH", "READ", "SCIE".}

\item{groups}{Name of one or more factors used for grouping.}

\item{final_weight}{Name of a column that contains final student weights.}

\item{data}{Name of a data frame containing columns given in previous arguments.}

\item{se}{TRUE if standard error are to be returned, FALSE if only means are to be returned.}

\item{brr_weights}{Names of the columns that contain BRR weights. If se is set to FALSE,
this argument is redunant and thus the default value is "".}
}
\value{
Data frame with columns for each given factor and student's average performances with standard errors.
}
\description{
Average student's performances with standard errors by given factor variables.
}

