% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpv.R
\name{mean_pvse}
\alias{mean_pvse}
\title{Average performance for one plausible value and a given replicate weight.}
\usage{
mean_pvse(data, pvname, groups, final_weights)
}
\arguments{
\item{data}{Name of a data frame containing columns given in other arguments.}

\item{pvname}{Name of the plausible value column.}

\item{groups}{Name of one or more factors used for grouping.}

\item{final_weights}{Name of a column that contains final student weights.}
}
\value{
Data frame containing columns for each of given factor variables, five columns with
        means calculated for each plausible values and a column with sums of weights.
}
\description{
Average performance for one plausible value and a given replicate weight.
}

