% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicetext.R
\name{nice_text}
\alias{nice_text}
\title{Changing distances between student's performances to make rainbow plot more readable.}
\usage{
nice_text(ave_perfs)
}
\arguments{
\item{ave_perfs}{Vector of average performances filtered by plot_rainbow function.}
}
\value{
Modified vector in which distance between adjacent elements are greater then 8.
}
\description{
Changing distances between student's performances to make rainbow plot more readable.
}

