% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdotggvis.R
\name{plot_dot}
\alias{plot_dot}
\title{Dot plot of average performances of all countries in given year.}
\usage{
plot_dot(csubject, cyear, isco_cats = as.character(1:9))
}
\arguments{
\item{csubject}{Character of the from "MATH"/"READ"/"SCIE".}

\item{cyear}{Chosen years as a character.}

\item{isco_cats}{ISCO categories to display.}
}
\value{
ggvis plot.
}
\description{
Dot plot of average performances of all countries in given year.
}

