% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbow.R
\name{plot_rainbow}
\alias{plot_rainbow}
\title{Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.}
\usage{
plot_rainbow(csubject, cnts, cyear)
}
\arguments{
\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{cyear}{Chosen year as a character - 2003/2006/2009/2012.}

\item{subject}{Character of the from "MATH"/"READ"/"SCIE".}
}
\value{
GGplot2 object.
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.
}

