% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbowshiny.R
\name{plot_rainbow_shiny}
\alias{plot_rainbow_shiny}
\title{Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.}
\usage{
plot_rainbow_shiny(cnts, cyear)
}
\arguments{
\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{cyear}{Chosen year as a character - 2003/2006/2009/2012.}
}
\value{
GGplot2 object.
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.
}

