% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbowtimeggvis.R
\name{plot_rainbow_time}
\alias{plot_rainbow_time}
\title{Rainbow plot of student's average performance over years.}
\usage{
plot_rainbow_time(csubject, ccnt)
}
\arguments{
\item{csubject}{Character of the from "MATH"/"READ"/"SCIE".}

\item{ccnt}{Chosen country.}
}
\value{
ggvis plot.
}
\description{
Rainbow plot of student's average performance over years.
}

