% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspreadallshiny.R
\name{plot_spread_all_shiny}
\alias{plot_spread_all_shiny}
\title{Plot of countries and all ISCO categories student's average performances by year designed for Shiny App.}
\usage{
plot_spread_all_shiny(csubject, ccnts)
}
\arguments{
\item{csubject}{Character of the from "MATH"/"READ"/"SCIE".}

\item{ccnts}{Country codes of two countries to compare.}
}
\value{
GGplot2 object.
}
\description{
Plot of countries and all ISCO categories student's average performances by year designed for Shiny App.
}

