% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottimeshiny.R
\name{plot_time_shiny}
\alias{plot_time_shiny}
\title{Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.}
\usage{
plot_time_shiny(csubject, cnts, isco_cats, years)
}
\arguments{
\item{csubject}{Character of the from "MATH"/"READ"/"SCIE".}

\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{isco_cats}{ISCO categories to plot - as a character - numbers of categories.}

\item{years}{Chosen years  - as a character - 2003/2006/2009/2012.}
}
\value{
GGplot2 object.
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.
}

