% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdot.R
\name{plotDot}
\alias{plotDot}
\title{Dot plot of average performances of all countries in given year.}
\usage{
plotDot(csubject, cyear, isco_cats)
}
\arguments{
\item{csubject}{Chosen subject - character "MATH"/"READ"/"SCIE"}

\item{cyear}{Chosen year of PISA study.}

\item{isco_cats}{Chosen isco categories.}
}
\value{
ggplot2 plot.
}
\description{
Dot plot of average performances of all countries in given year.
}

