% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbowtime.R
\name{plotRainbowTime}
\alias{plotRainbowTime}
\title{Rainbow plot of student's average performance over years.}
\usage{
plotRainbowTime(csubject, cnts)
}
\arguments{
\item{sdf}{Data frame returned from reactive function.}
}
\value{
ggplot2 plot.
}
\description{
Rainbow plot of student's average performance over years.
}

