% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilepv.R
\name{quantile_pv}
\alias{quantile_pv}
\title{Quantile of student's performances.}
\usage{
quantile_pv(pvname, prob, groups, data)
}
\arguments{
\item{pvname}{Character of the from "MATH"/"READ"/"SCIE".}

\item{prob}{Probability.}

\item{groups}{Characters with grouping variables names.}

\item{data}{Data frame that contains columns given in groups and pvname arguments.}
}
\value{
Data frame containing levels of given factors and a column with calculated values.
}
\description{
Quantile of student's performances.
}

