% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilespv.R
\name{quantiles_pv}
\alias{quantiles_pv}
\title{Quantiles of student's performances.}
\usage{
quantiles_pv(pvname, probs, groups, data)
}
\arguments{
\item{pvname}{Character of the from "MATH"/"READ"/"SCIE".}

\item{probs}{Vector of probabilities.}

\item{groups}{Characters with grouping variables names.}

\item{data}{Data frame that contains columns given in groups and pvname arguments.}
}
\value{
Data frame containing levels of given factors and a column for each calculated quantile.
}
\description{
Quantiles of student's performances.
}

