% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{spreadBySubject}
\alias{spreadBySubject}
\title{Helper function for exportSpreadsheet to convert data in wide format.}
\usage{
spreadBySubject(sourceTibble, column)
}
\arguments{
\item{sourceTibble}{tibble returned by importFromTxt function.}

\item{column}{name of a column to spread.}
}
\value{
tibble
}
\description{
Helper function for exportSpreadsheet to convert data in wide format.
}

