% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torainbow.R
\name{to_rainbow}
\alias{to_rainbow}
\title{Preparing data to plot.}
\usage{
to_rainbow(subject, cnt_col, cnts, isco_cols, final_weight, data)
}
\arguments{
\item{subject}{String of the form "MATH"/"READ"/"SCIE".}

\item{cnt_col}{Names of the columns that contain country codes.}

\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{isco_cols}{Names of the columns with father's and mother's isco codes.}

\item{final_weight}{Name of a column with final student weights.}

\item{data}{Data frame returned by mean_by_all_cats function.}
}
\value{
Data frame with column for country code, number that is used by plot_rainbow function
        for x axis, isco categories along with country average perf. and means for each category.
}
\description{
Preparing data to plot.
}

