% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Imputation variance for given set of plausible values and weights.}
\usage{
var_imp(means_ppv)
}
\arguments{
\item{means_ppv}{Data frame returned by mean_ppvs function that contains average performances
computed for each plausible value.}
}
\value{
Vector containing computed imputation variances.
}
\description{
Imputation variance for given set of plausible values and weights.
}

