% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsmlopv.R
\name{var_sml_opv}
\alias{var_sml_opv}
\title{Sampling variance for one plausible values and a set of replicate weights.}
\usage{
var_sml_opv(pvname, groups, brr_weights, means, data)
}
\arguments{
\item{pvname}{String of the form "MATH", "READ", "SCIE".}

\item{groups}{Name of one or more factors used for grouping.}

\item{brr_weights}{Names of columns that contain BRR weights.}

\item{means}{Vector of average performances computed by mean_o function.}

\item{data}{Name of a data frame containing columns given in previous arguments.}
}
\value{
Vector that contains computed sampling variances.
}
\description{
Sampling variance for one plausible values and a set of replicate weights.
}

