% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscocat.R
\name{isco_cat}
\alias{isco_cat}
\title{Extract first and first two digits of ISCO code as factors both for mother's and father's occupation.}
\usage{
isco_cat(data, isco_cols)
}
\arguments{
\item{data}{Name of a data frame containing columns given in isco_cols argument.}

\item{isco_cols}{Names of columns with mother's and father's occupation codes in this order.}
}
\value{
Data frame given in 'data' argument with four additional columns for mother's and father's
        primary and secondary ISCO categories.
}
\description{
Extract first and first two digits of ISCO code as factors both for mother's and father's occupation.
}

