% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdotggvis.R
\name{plot_dot}
\alias{plot_dot}
\title{Dot plot of average performances of all countries in given year.}
\usage{
plot_dot(sdf)
}
\arguments{
\item{sdf}{Data frame returned by interactive() function after filtering data.}
}
\value{
ggvis plot.
}
\description{
Dot plot of average performances of all countries in given year.
}

