% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowplot.R
\name{plot_rainbow}
\alias{plot_rainbow}
\title{Rainbow plot of average performances for two countries grouped by isco categories.}
\usage{
plot_rainbow(subject, cnt_col, cnts, isco_cols, final_weight, data)
}
\arguments{
\item{subject}{String of the form "MATH"/"READ"/"SCIE".}

\item{cnt_col}{Names of the columns that contain country codes.}

\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{isco_cols}{Names of the columns with father's and mother's isco codes.}

\item{final_weight}{Name of a column with final student weights.}

\item{data}{Data frame returned by mean_by_all_cats function.}
}
\value{
GGplot2 object.
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories.
}

