% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrainbowtime.R
\name{plot_rainbow_time}
\alias{plot_rainbow_time}
\title{Rainbow plot of student's average performance over years.}
\usage{
plot_rainbow_time(sdf)
}
\arguments{
\item{sdf}{Data frame returned from reactive function.}
}
\value{
ggvis plot.
}
\description{
Rainbow plot of student's average performance over years.
}

