% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottime.R
\name{plot_time_shiny}
\alias{plot_time_shiny}
\title{Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.}
\usage{
plot_time_shiny(csubject, cnts, isco_cats = as.character(1:9))
}
\arguments{
\item{csubject}{Character of the from "MATH"/"READ"/"SCIE".}

\item{cnts}{Country codes of countries to compare on a rainbow plot.}

\item{isco_cats}{ISCO categories to plot - as a character - numbers of categories.}
}
\value{
GGplot2 object.
}
\description{
Rainbow plot of average performances for two countries grouped by isco categories designed for shiny app.
}

