% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepv.R
\name{se_pv}
\alias{se_pv}
\title{Standard errors for average performances grouped by given factors.}
\usage{
se_pv(pvname, groups, final_weight, brr_weights, means_ppv, data)
}
\arguments{
\item{pvname}{String of the form "MATH", "READ", "SCIE".}

\item{groups}{Name of one or more factors used for grouping.}

\item{final_weight}{Name of a column that contains final student weights.}

\item{brr_weights}{Names of the columns that contain BRR weights.}

\item{means_ppv}{Data frame returned by mean_ppvs function that contains average performances
computed for each plausible value.}

\item{data}{Name of a data frame containing columns given in previous arguments.}
}
\value{
Vector that contains computed standard errors.
}
\description{
Standard errors for average performances grouped by given factors.
}

