% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsml.R
\name{var_sml}
\alias{var_sml}
\title{Sampling variance for a given set of plausible values and weights.}
\usage{
var_sml(pvname, groups, brr_weights, final_weight, data)
}
\arguments{
\item{pvname}{String of the form "MATH", "READ", "SCIE".}

\item{groups}{Name of one or more factors used for grouping.}

\item{brr_weights}{Names of columns that contain BRR weights.}

\item{data}{Name of a data frame containing columns given in previous arguments.}

\item{means}{Vector of average performances computed by mean_o function.}
}
\value{
Vector that contains computed sampling variances.
}
\description{
Sampling variance for a given set of plausible values and weights.
}

